changes until now.
Adjusting the file.
vector extension works just before version 14. 
So I calculate vector using the Euclidian distance function from numpy. For now, I did a cross-join between the values on the same column of the file imported with the embeddings. 
the vector comparison between the table and the text still is not working. -> Similarity Vector Search with the results